Create Exercise_3.asm
Given a string of doublewords S, write an assembly program that creates a string D1 by selecting the higher words from the doubleword elements in S, which are prime numbers. The program also creates a string D2 by selecting those words from the doublewords in S, which in the hexadecimal representation contain no letters. The program computes the maximum value of string D1 and prints on the screen how many times the max appears in the string D2.