Rename Ecercise_4.asm to Exercise_4.asm
Write a program that reads a doubleword N from the standard input device and then it reads  a string of doubleword numbers from the standard input device. The program will store these numbers in a doubleword. Then it will build a new string of bytes which contains, for each doubleword from the first string, the sum of the even digits (cifrele pare) of the representation in base 10 of this doubleword.